src/store/features/counter/counterSlice.ts
import { createSlice, PayloadAction } from '@reduxjs/toolkit';

export type CounterState = {
  value: number;
};

const initialState: CounterState = {
  value: 0,
};

export const counterSlice = createSlice({
  name: 'counter',
  initialState,
  /**
   * RTK uses Immer internally, so you can "mutate" the state as it's only a draft.
   * https://redux-toolkit.js.org/usage/immer-reducers#redux-toolkit-and-immer
   */
  reducers: {
    increment: state => {
      state.value++;
    },
    decrement: state => {
      state.value--;
    },
    incrementByAmount: (state, action: PayloadAction<number>) => {
      state.value += action.payload;
    },
  },
});

export const counterActions = counterSlice.actions;

export const counterReducer = counterSlice.reducer;
